Check arguments for initialize_resource (#111)
* Check arguments for initialize_resource

Here we refactor argument checks for resource initialization out into static methods.
The static methods also perform type casting to the desired type.

A potential bug was discovered, where permissions were being downcasted from long to int.
This has been repaired, and signatures have been adjusted to use long instead of int where required.
The type conversion is now happening in exactly one place.

* Style fixes

+++ ext/semian/semian.h
@@ -47,10 +47,10 @@ key_t
 generate_key(const char *name);
 
 void
-set_semaphore_permissions(int sem_id, int permissions);
+set_semaphore_permissions(int sem_id, long permissions);
 
 int
-create_semaphore(int key, int permissions, int *created);
+create_semaphore(int key, long permissions, int *created);
 
 int
 get_semaphore(int key);
